backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
email-validator>=2.0.0
motor>=3.4.0
pymongo>=4.7.0
requests>=2.31.0
trafilatura>=1.9.0
python-dotenv>=1.0.0
bcrypt>=4.1.0
python-jose[cryptography]>=3.3.0
gTTS>=2.5.0
pydub>=0.25.0
imageio-ffmpeg>=0.6.0
python-multipart>=0.0.9
pyaudioop>=0.3.0
certifi>=2024.2.2
